Replaced the HANDLE_SESSION_ERROR macro with CHECK_RETURN_CODE and CHECK_RETURN_POINTER.
Both macros call the RAISE_SSH2_ERROR macro (also new), that can also be used to
raise a libssh2.Error for the last session error without checking for the return
value. That might be useful when you need to do some cleanup before leaving the
function in case of an error.