Removed of PyCObject code, that provided an API for other C extension modules.
The code was not tested at least since version 0.6.9, no header file in order
to use the API when writing C extensions is provided and it would need some
compatibility changes for Python 3.2. The libssh2 module is currently completly
broken on Python 3.2, due to that code.

Because of those issues I have decided to no longer maintain that code.